feat: 閉合 JOIN 裝備重算交易

+++ remake/internal/campaign/native_shop_equipment_recipient.go
@@ -91,10 +91,10 @@ func NativeShopEquipmentCurrentStats(record []byte) ([4]int, error) {
 		)
 	}
 	return [4]int{
-		int(binary.LittleEndian.Uint16(record[0x48:])),
-		int(binary.LittleEndian.Uint16(record[0x4a:])),
-		int(binary.LittleEndian.Uint16(record[0x4c:])),
-		int(binary.LittleEndian.Uint16(record[0x4e:])),
+		int(int16(binary.LittleEndian.Uint16(record[0x48:]))),
+		int(int16(binary.LittleEndian.Uint16(record[0x4a:]))),
+		int(int16(binary.LittleEndian.Uint16(record[0x4c:]))),
+		int(int16(binary.LittleEndian.Uint16(record[0x4e:]))),
 	}, nil
 }
